# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Recipes",
    "description": "Site where people can shares recipes"
  },
  "definitions": {
    "ApiError": {
      "properties": {
        "message": {
          "type": "string",
          "example": "error text"
        },
        "statusCode": {
          "type": "number",
          "example": 400
        }
      }
    },
    "MongoId": {
      "type": "string",
      "example": "63db2a91751809b97f5a7e12"
    },
...
  },
  "paths": {
...
    "/authors/userName": {
      "patch": {
        "description": "change userName",
        "tags": [
          "Author"
        ],
        "operationId": "UserName",
        "parameters": [
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          },
          {
            "name": "userName",
            "type": "string",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "properties": {
                "message": {
                  "type": "string",
                  "example": "userName has been changed to new_userName"
                },
                "statusCode": {
                  "type": "number",
                  "example": 200
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      }
    },
...
    "/authors/subscribers": {
      "get": {
        "description": "Get author's subscribers",
        "tags": [
          "Author"
        ],
        "operationId": "GetSubscribers",
        "parameters": [
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "properties": {
                "_id": {
                  "$ref": "#/definitions/MongoId"
                },
                "userName": {
                  "type": "string",
                  "example": "john_01"
                },
                "email": {
                  "type": "string",
                  "example": "john1999@mail.com"
                },
                "avatar": {
                  "type": "string",
                  "example": "https://www.google.com/image"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      }
    },
...
    "/recipes/{recipeId}/book-toggle": {
      "patch": {
        "description": "Add or remove recipe from your book",
        "tags": [
          "Recipe"
        ],
        "operationId": "BookToggle",
        "parameters": [
          {
            "name": "recipeId",
            "type": "string",
            "in": "path",
            "required": true
          },
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "properties": {
                "message": {
                  "type": "string",
                  "example": "recipe has been added to book"
                },
                "statusCode": {
                  "type": "number",
                  "example": 200
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      }
    },
...
    "/recipes/reviews": {
      "get": {
        "description": "Get reviews by recipe id",
        "tags": [
          "Recipe"
        ],
        "operationId": "GetReviews",
        "parameters": [
          {
            "name": "recipeId",
            "type": "string",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "properties": {
                "_id": {
                  "$ref": "#/definitions/MongoId"
                },
                "text": {
                  "type": "string",
                  "example": "review text"
                },
                "photo": {
                  "type": "string",
                  "example": "reviews/63e0f76a03a811f4473b0664/1675687786936.jpg"
                },
                "owner": {
                  "properties": {
                    "userName": {
                      "type": "string",
                      "example": "user_name"
                    },
                    "avatar": {
                      "type": "string",
                      "example": "authors/63e0f3ab03a811f4473b05dc/1676022066729.jpg"
                    }
                  }
                },
                "createdAt": {
                  "type": "string",
                  "example": "2023-02-06T12:49:46.923Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      }
    },
    "/recipes/{recipeId}/addReview": {
      "post": {
        "description": "Add review to recipe by id",
        "tags": [
          "Recipe"
        ],
        "operationId": "AddReview",
        "parameters": [
          {
            "name": "recipeId",
            "type": "string",
            "in": "path",
            "required": true
          },
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          },
          {
            "name": "photo",
            "type": "file",
            "in": "formData",
            "required": false
          },
          {
            "name": "text",
            "type": "string",
            "in": "formData",
            "required": true
          },
          {
            "name": "rating",
            "type": "number",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      }
    },
...
    "/reviews/{reviewId}": {
      "delete": {
        "description": "Delete review by id",
        "tags": [
          "Review"
        ],
        "operationId": "DeleteReview",
        "parameters": [
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/ApiError"
            }
          }
        }
      }
    }
  }
}